+++ src/PrintOrder/utils.test.js
@@ -3,7 +3,11 @@ import { getOrderPrintData } from './utils';
 
 describe('getOrderPrintData', () => {
   it('should call ky', () => {
-    const ky = { get: jest.fn() };
+    const ky = {
+      get: jest.fn(() => ({
+        json: () => Promise.resolve(),
+      })),
+    };
 
     getOrderPrintData(ky, order);